refactor(fieldmap): move suppress phenotype modal into plot details modal
